moved tables out of lookup
